Adding screen UI

+++ data_remote/src/main/java/com/z_company/entity/SectionElectric.kt
@@ -6,9 +6,9 @@ data class SectionElectric(
     var sectionId: String = "",
     var locoId: String = "",
     val type: LocoType = LocoType.ELECTRIC,
-    var acceptedEnergy: Int? = null,
-    var deliveryEnergy: Int? = null,
-    var acceptedRecovery: Int? = null,
-    var deliveryRecovery: Int? = null
+    var acceptedEnergy: Double? = null,
+    var deliveryEnergy: Double? = null,
+    var acceptedRecovery: Double? = null,
+    var deliveryRecovery: Double? = null
 )
